requirements.txt
pymongo>=4.5.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-community>=0.0.20
pandas>=2.0.0
pydantic>=2.0.0
rich>=13.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
faker>=19.0.0